pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.12",
    "fastapi>=0.115.9",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "numpy>=2.3.0",
    "openai>=1.88.0",
    "pinecone-client>=6.0.0",
    "pypdf>=5.6.0",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "tiktoken>=0.9.0",
    "unstructured>=0.17.2",
    "uvicorn>=0.34.3",
]
